Prediccion con el modelo entrenado

+++ alembic/versions/415d144649b5_crear_tabla_libros.py
@@ -25,7 +25,7 @@ def upgrade() -> None:
     sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
     sa.Column('titulo', sa.String(length=255), nullable=False),
     sa.Column('autor', sa.String(length=255), nullable=False),
-    sa.Column('rating', sa.Integer(), nullable=False),
+    sa.Column('rating', sa.Float(), nullable=False),
     sa.PrimaryKeyConstraint('id')
     )
     # ### end Alembic commands ###
